chore(core): refactor studio backend
