+++ src/bp/core/app/server.ts
@@ -7,11 +7,12 @@ import { RequestWithUser } from 'common/typings'
 import compression from 'compression'
 import cookieParser from 'cookie-parser'
 import session from 'cookie-session'
-import { BotService } from 'core/bots'
+import { TYPES } from 'core/app/types'
+import { BotService, BotsRouter } from 'core/bots'
 import { GhostService } from 'core/bpfs'
-import { CMSRouter, CMSService } from 'core/cms'
+import { CMSService } from 'core/cms'
 import { ExternalAuthConfig, ConfigProvider } from 'core/config'
-import { ConverseRouter, ConverseService } from 'core/converse'
+import { ConverseService } from 'core/converse'
 import { FlowService, SkillService } from 'core/dialog'
 import { JobService } from 'core/distributed'
 import { AlertingService, MonitoringService } from 'core/health'
@@ -31,16 +32,13 @@ import {
   TOKEN_AUDIENCE
 } from 'core/security'
 import { TelemetryRouter, TelemetryRepository } from 'core/telemetry'
-import { ActionService, ActionServersService, HintsService, HintsRouter } from 'core/user-code'
+import { ActionService, ActionServersService, HintsService } from 'core/user-code'
 import { WorkspaceService } from 'core/users'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
 import { UnlicensedError } from 'errors'
 import express, { NextFunction, Response } from 'express'
 import rateLimit from 'express-rate-limit'
-import { Request } from 'express-serve-static-core'
-import rewrite from 'express-urlrewrite'
-import fs from 'fs'
 import { createServer, Server } from 'http'
 import { inject, injectable, postConstruct, tagged } from 'inversify'
 import jsonwebtoken from 'jsonwebtoken'
@@ -51,32 +49,18 @@ import { Memoize } from 'lodash-decorators'
 import ms from 'ms'
 import path from 'path'
 import portFinder from 'portfinder'
+import { StudioRouter } from 'studio/studio-router'
 import { URL } from 'url'
 import yn from 'yn'
 
-import { BotsRouter } from '../routers'
-import { NLURouter } from '../routers/bots/nlu'
 import { isDisabled } from '../routers/conditionalMiddleware'
 import { SdkApiRouter } from '../routers/sdk/router'
 import { ShortLinksRouter } from '../routers/shortlinks'
 import { NLUService } from '../services/nlu/nlu-service'
-import { TYPES } from './types'
+import { debugRequestMw, resolveAsset } from './server_utils'
 
 const BASE_API_PATH = '/api/v1'
 const SERVER_USER_STRATEGY = 'default' // The strategy isn't validated for the userver user, it could be anything.
-
-const debug = DEBUG('api')
-const debugRequest = debug.sub('request')
-
-const debugRequestMw = (req: Request, _res, next) => {
-  debugRequest(`${req.path} %o`, {
-    method: req.method,
-    ip: req.ip,
-    originalUrl: req.originalUrl
-  })
-
-  next()
-}
 
 @injectable()
 export class HTTPServer {
@@ -86,12 +70,9 @@ export class HTTPServer {
 
   private readonly adminRouter: AdminRouter
   private readonly botsRouter: BotsRouter
-  private cmsRouter!: CMSRouter
-  private nluRouter!: NLURouter
+  private readonly studioRouter!: StudioRouter
   private readonly modulesRouter: ModulesRouter
   private readonly shortLinksRouter: ShortLinksRouter
-  private converseRouter!: ConverseRouter
-  private hintsRouter!: HintsRouter
   private telemetryRouter!: TelemetryRouter
   private mediaRouter: MediaRouter
   private readonly sdkApiRouter!: SdkApiRouter
@@ -105,7 +86,6 @@ export class HTTPServer {
     resource: string,
     noAudit?: boolean
   ) => Promise<boolean>
-  private indexCache: { [pageUrl: string]: string } = {}
 
   private jwksClient?: jwksRsa.JwksClient
   private jwksKeyId?: string
@@ -126,17 +106,17 @@ export class HTTPServer {
     @inject(TYPES.NotificationsService) notificationService: NotificationsService,
     @inject(TYPES.SkillService) skillService: SkillService,
     @inject(TYPES.GhostService) private ghostService: GhostService,
-    @inject(TYPES.HintsService) private hintsService: HintsService,
+    @inject(TYPES.HintsService) hintsService: HintsService,
     @inject(TYPES.LicensingService) licenseService: LicensingService,
-    @inject(TYPES.ConverseService) private converseService: ConverseService,
+    @inject(TYPES.ConverseService) converseService: ConverseService,
     @inject(TYPES.WorkspaceService) private workspaceService: WorkspaceService,
     @inject(TYPES.BotService) private botService: BotService,
-    @inject(TYPES.AuthStrategies) private authStrategies: AuthStrategies,
+    @inject(TYPES.AuthStrategies) authStrategies: AuthStrategies,
     @inject(TYPES.MonitoringService) private monitoringService: MonitoringService,
     @inject(TYPES.AlertingService) private alertingService: AlertingService,
     @inject(TYPES.JobService) private jobService: JobService,
     @inject(TYPES.LogsRepository) private logsRepo: LogsRepository,
-    @inject(TYPES.NLUService) private nluService: NLUService,
+    @inject(TYPES.NLUService) nluService: NLUService,
     @inject(TYPES.TelemetryRepository) private telemetryRepo: TelemetryRepository
   ) {
     this.app = express()
@@ -179,22 +159,35 @@ export class HTTPServer {
       this.logsRepo,
       authStrategies
     )
-    this.shortLinksRouter = new ShortLinksRouter(this.logger)
-    this.botsRouter = new BotsRouter({
-      actionService,
-      actionServersService,
-      botService,
-      cmsService,
-      configProvider,
-      flowService,
-      mediaServiceProvider,
-      logsService,
-      notificationService,
+
+    this.studioRouter = new StudioRouter(
+      logger,
       authService,
       workspaceService,
-      moduleLoader,
-      logger: this.logger
-    })
+      botService,
+      configProvider,
+      actionService,
+      cmsService,
+      flowService,
+      notificationService,
+      logsService,
+      ghostService,
+      mediaServiceProvider,
+      actionServersService
+    )
+
+    this.shortLinksRouter = new ShortLinksRouter(this.logger)
+    this.botsRouter = new BotsRouter(
+      botService,
+      configProvider,
+      authService,
+      workspaceService,
+      nluService,
+      converseService,
+      hintsService,
+      this.logger,
+      this
+    )
     this.sdkApiRouter = new SdkApiRouter(this.logger)
     this.telemetryRouter = new TelemetryRouter(this.logger, this.authService, this.telemetryRepo)
     this.mediaRouter = new MediaRouter(
@@ -234,30 +227,12 @@ export class HTTPServer {
     this.httpServer = createServer(app)
 
     await this.mediaRouter.initialize()
-    await this.botsRouter.initialize()
-    this.cmsRouter = new CMSRouter(
-      this.logger,
-      this.authService,
-      this.cmsService,
-      this.workspaceService,
-      this.ghostService
-    )
-    this.nluRouter = new NLURouter(this.logger, this.authService, this.workspaceService, this.nluService)
-    this.converseRouter = new ConverseRouter(this.logger, this.converseService, this.authService, this)
-    this.hintsRouter = new HintsRouter(this.logger, this.hintsService, this.authService, this.workspaceService)
-    this.botsRouter.router.use('/content', this.cmsRouter.router)
-    this.botsRouter.router.use('/converse', this.converseRouter.router)
-    this.botsRouter.router.use('/nlu', this.nluRouter.router)
 
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     AppLifecycle.waitFor(AppLifecycleEvents.BOTPRESS_READY).then(() => {
       this.isBotpressReady = true
     })
-
-    this.botsRouter.router.use('/hints', this.hintsRouter.router)
   }
-
-  resolveAsset = file => path.resolve(process.PROJECT_LOCATION, 'data/assets', file)
 
   async start() {
     const botpressConfig = await this.configProvider.getBotpressConfig()
@@ -361,13 +336,14 @@ export class HTTPServer {
       `)
     })
 
-    this.app.use('/assets', this.guardWhiteLabel(), express.static(this.resolveAsset('')))
-    this.app.use(rewrite('/:app/:botId/*env.js', '/api/v1/bots/:botId/:app/js/env.js'))
-
     this.adminRouter.setupRoutes(this.app)
+    await this.botsRouter.setupRoutes(this.app)
+    await this.studioRouter.setupRoutes(this.app)
+
+    this.app.use('/assets', this.guardWhiteLabel(), express.static(resolveAsset('')))
 
     this.app.use(`${BASE_API_PATH}/modules`, this.modulesRouter.router)
-    this.app.use(`${BASE_API_PATH}/bots/:botId`, this.botsRouter.router)
+
     this.app.use(`${BASE_API_PATH}/sdk`, this.sdkApiRouter.router)
     this.app.use(`${BASE_API_PATH}/telemetry`, this.telemetryRouter.router)
     this.app.use(`${BASE_API_PATH}/media`, this.mediaRouter.router)
@@ -403,8 +379,6 @@ export class HTTPServer {
       })
     })
 
-    this.setupStaticRoutes(this.app)
-
     process.HOST = config.host
     process.PORT = await portFinder.getPortPromise({ port: config.port })
     process.EXTERNAL_URL = process.env.EXTERNAL_URL || config.externalUrl || `http://${process.HOST}:${process.PORT}`
@@ -435,49 +409,6 @@ export class HTTPServer {
       }
       next()
     }
-  }
-
-  setupStaticRoutes(app) {
-    // Dynamically updates the static paths of index files
-    const resolveIndexPaths = page => (req, res) => {
-      res.contentType('text/html')
-
-      // Not caching pages in dev (issue with webpack )
-      if (this.indexCache[page] && process.IS_PRODUCTION) {
-        return res.send(this.indexCache[page])
-      }
-
-      fs.readFile(this.resolveAsset(page), (err, data) => {
-        if (data) {
-          this.indexCache[page] = data
-            .toString()
-            .replace(/\<base href=\"\/\" ?\/\>/, `<base href="${process.ROOT_PATH}/" />`)
-            .replace(/ROOT_PATH=""|ROOT_PATH = ''/, `window.ROOT_PATH="${process.ROOT_PATH}"`)
-
-          res.send(this.indexCache[page])
-        } else {
-          res.sendStatus(404)
-        }
-      })
-    }
-
-    app.get('/studio', (req, res, next) => res.redirect('/admin'))
-
-    app.use('/:app(studio)/:botId', express.static(this.resolveAsset('ui-studio/public'), { index: false }))
-    app.use('/:app(studio)/:botId', resolveIndexPaths('ui-studio/public/index.html'))
-
-    app.use('/:app(lite)/:botId?', express.static(this.resolveAsset('ui-studio/public/lite'), { index: false }))
-    app.use('/:app(lite)/:botId?', resolveIndexPaths('ui-studio/public/lite/index.html'))
-
-    app.use('/:app(lite)/:botId', express.static(this.resolveAsset('ui-studio/public'), { index: false }))
-    app.use('/:app(lite)/:botId', resolveIndexPaths('ui-studio/public/index.html'))
-
-    app.get(['/:app(studio)/:botId/*'], resolveIndexPaths('ui-studio/public/index.html'))
-
-    app.use('/admin', express.static(this.resolveAsset('admin/ui/public'), { index: false }))
-    app.get(['/admin', '/admin/*'], resolveIndexPaths('admin/ui/public/index.html'))
-
-    app.get('/', (req, res) => res.redirect(`${process.ROOT_PATH}/admin`))
   }
 
   createRouterForBot(router: string, identity: string, options: RouterOptions): any & http.RouterExtension {
