+++ src/bp/core/bots/index.ts
@@ -1,2 +1,3 @@
 export * from './bot-monitoring-service'
 export * from './bot-service'
+export * from './bots-router'
