+++ src/bp/core/cms/index.ts
@@ -1,4 +1,3 @@
-export * from './cms-router'
 export * from './cms-service'
 export * from './templating'
 export * from './render-service'
