+++ src/bp/core/modules/modules-router.ts
@@ -73,5 +73,14 @@ export class ModulesRouter extends CustomRouter {
         res.send(await this.moduleLoader.getTranslations())
       })
     )
+
+    this.router.get(
+      '/dialogConditions',
+      this.checkTokenHeader,
+      this.asyncMiddleware(async (req, res) => {
+        const conditions = await this.moduleLoader.getDialogConditions()
+        res.send(conditions)
+      })
+    )
   }
 }
