+++ src/bp/core/routers/index.ts
@@ -1,4 +1,3 @@
-export * from './bots'
 export * from './util'
 export * from './errors'
 export * from './conditionalMiddleware'
