+++ src/bp/core/security/router-security.ts
@@ -1,5 +1,6 @@
 import { checkRule, CSRF_TOKEN_HEADER_LC, JWT_COOKIE_NAME } from 'common/auth'
 import { RequestWithUser } from 'common/typings'
+import { ConfigProvider } from 'core/config'
 import {
   InvalidOperationError,
   BadRequestError,
@@ -10,7 +11,7 @@ import {
   UnauthorizedError
 } from 'core/routers'
 import { WorkspaceService } from 'core/users'
-import { NextFunction, Request, Response } from 'express'
+import { NextFunction, Request, RequestHandler, Response } from 'express'
 import { AuthService, WORKSPACE_HEADER, SERVER_USER } from './auth-service'
 
 const debugFailure = DEBUG('audit:collab:fail')
@@ -217,4 +218,38 @@ const checkPermissions = (workspaceService: WorkspaceService) => (
   }
 
   audit(debugSuccess, { userRole: role?.id })
+}
+
+const mediaPathRegex = new RegExp(/^\/api\/v(\d)\/bots\/[A-Z0-9_-]+\/media\//, 'i')
+export const checkBotVisibility = (configProvider: ConfigProvider, checkTokenHeader: RequestHandler) => async (
+  req,
+  res,
+  next
+) => {
+  // '___' is a non-valid botId, but here acts as for "all bots"
+  // This is used in modules when they setup routes that work on a global level (they are not tied to a specific bot)
+  // Check the 'sso-login' module for an example
+  if (req.params.botId === '___' || req.originalUrl.endsWith('env.js')) {
+    return next()
+  }
+
+  try {
+    const config = await configProvider.getBotConfig(req.params.botId)
+    if (config.disabled) {
+      // The user must be able to get the config to change the bot status
+      if (req.originalUrl.endsWith(`/api/v1/bots/${req.params.botId}`)) {
+        return next()
+      }
+
+      return next(new NotFoundError('Bot is disabled'))
+    }
+
+    if (config.private && !mediaPathRegex.test(req.originalUrl)) {
+      return checkTokenHeader(req, res, next)
+    }
+  } catch (err) {
+    return next(new NotFoundError('Invalid Bot ID'))
+  }
+
+  next()
 }
