+++ src/bp/studio/utils/custom-studio-router.ts
@@ -1,0 +1,62 @@
+import { Logger } from 'botpress/sdk'
+import { AsyncMiddleware, asyncMiddleware } from 'common/http'
+import { RequestWithUser } from 'common/typings'
+import { BotService } from 'core/bots'
+import { GhostService } from 'core/bpfs'
+import { CMSService } from 'core/cms'
+import { ConfigProvider } from 'core/config/config-loader'
+import { FlowService } from 'core/dialog'
+import { LogsService } from 'core/logger'
+import { MediaServiceProvider } from 'core/media'
+import { NotificationsService } from 'core/notifications'
+import { AuthService, TOKEN_AUDIENCE, needPermissions, hasPermissions, checkTokenHeader } from 'core/security'
+import { ActionServersService, ActionService } from 'core/user-code'
+import { WorkspaceService } from 'core/users'
+import { RequestHandler, Router } from 'express'
+import { StudioServices } from 'studio/studio-router'
+
+export abstract class CustomStudioRouter {
+  protected logger: Logger
+  protected authService: AuthService
+  protected configProvider: ConfigProvider
+  protected botService: BotService
+  protected mediaServiceProvider: MediaServiceProvider
+  protected cmsService: CMSService
+  protected workspaceService: WorkspaceService
+  protected logsService: LogsService
+  protected flowService: FlowService
+  protected actionService: ActionService
+  protected actionServersService: ActionServersService
+  protected notificationService: NotificationsService
+  protected bpfs: GhostService
+
+  protected readonly needPermissions: (operation: string, resource: string) => RequestHandler
+  protected readonly asyncMiddleware: AsyncMiddleware
+  protected readonly hasPermissions: (req: RequestWithUser, operation: string, resource: string) => Promise<boolean>
+  protected readonly checkTokenHeader: RequestHandler
+
+  public readonly router: Router
+
+  constructor(name: string, services: StudioServices) {
+    this.asyncMiddleware = asyncMiddleware(services.logger, name)
+    this.needPermissions = needPermissions(services.workspaceService)
+    this.hasPermissions = hasPermissions(services.workspaceService)
+    this.checkTokenHeader = checkTokenHeader(services.authService, TOKEN_AUDIENCE)
+
+    this.router = Router({ mergeParams: true })
+
+    this.logger = services.logger
+    this.authService = services.authService
+    this.configProvider = services.configProvider
+    this.botService = services.botService
+    this.mediaServiceProvider = services.mediaServiceProvider
+    this.workspaceService = services.workspaceService
+    this.cmsService = services.cmsService
+    this.flowService = services.flowService
+    this.actionService = services.actionService
+    this.actionServersService = services.actionServersService
+    this.notificationService = services.notificationService
+    this.logsService = services.logsService
+    this.bpfs = services.bpfs
+  }
+}
